requirements.txt
streamlit==1.27.2
python-dotenv==1.0.0
langchain==0.0.321
faiss-cpu==1.7.4
pymongo==4.5.0
transformers==4.34.1
torch==2.1.0
pdf2image==1.16.3
google-generativeai==0.2.1
anthropic==0.3.11
openai==0.28.1
sentence-transformers==2.2.2
pypdf==3.16.1

#streamlit
#python-dotenv
#langchain
#faiss-cpu
#pymongo
#transformers
#torch
#pdf2image
#google-generativeai
#anthropic